Refactor weather forecast initialization logic
Consolidated initialization logic by removing `initializeForecastDataPerDay` and `initializeUnits` functions and integrating their functionality into `initializeWeatherForecast`. Simplified `WeatherForecastDataService` by removing `forecastPerDaySubject` and `units` properties and related methods. This reduces complexity and improves maintainability.
