--- src/DevServerClient/src/app/Services/weather-forecast/units-initializer.ts
@@ -1,15 +1,0 @@
-import { Units } from '../../models/weather-forecast/units';
-
-export function initializeUnits(): Units {
-    return {
-        predictability: '',
-        precipitation: '',
-        windSpeed: '',
-        precipitationProbability: '',
-        relativeHumidity: '',
-        temperature: '',
-        time: '',
-        pressure: '',
-        windDirection: ''
-    } as Units;
-}
