+++ src/DevServerClient/src/app/Services/weather-forecast/weather-forecast-data.service.ts
@@ -3,11 +3,7 @@ import { BehaviorSubject, Observable } from 'rxjs';
 import { LocationQueryResult } from '../../models/weather-forecast/locationQueryResult';
 import { WeatherForecast } from '../../models/weather-forecast/weatherForecast';
 import { initializeLocationQueryResult } from './locationQueryResult-initializer';
-import { initializeWeatherForecast } from './weather-forecast-initializer';
-import { initializeForecastDataPerDay } from './forecastPerDay-initializer';
-import { ForecastDataPerDay } from '../../models/weather-forecast/forecastDataPerDay';
-import { initializeUnits } from './units-initializer';
-import { Units } from '../../models/weather-forecast/units';
+import { initializeWeatherForecast } from './weatherForecast-initializer';
 
 @Injectable({
   providedIn: 'root'
@@ -15,13 +11,8 @@ import { Units } from '../../models/weather-forecast/units';
 export class WeatherForecastDataService {
   private locationQueryResultSubject = new BehaviorSubject<LocationQueryResult>(initializeLocationQueryResult());
   private weatherForecastSubject = new BehaviorSubject<WeatherForecast>(initializeWeatherForecast());
-  private forecastPerDaySubject = new BehaviorSubject<ForecastDataPerDay>(initializeForecastDataPerDay());
-  private units = new BehaviorSubject<Units>(initializeUnits());
 
-
-  constructor() {
-
-  }
+  constructor() { }
 
   setLocationQueryResult(data: LocationQueryResult) {
     this.locationQueryResultSubject.next(data);
@@ -38,8 +29,6 @@ export class WeatherForecastDataService {
 
   setWeatherForecast(data: WeatherForecast) {
     this.weatherForecastSubject.next(data);
-    this.forecastPerDaySubject.next(data.forecastDataPerDay);
-    this.units.next(data.units);
   }
 
   getWeatherForecast(): Observable<WeatherForecast | null> {
@@ -47,13 +36,5 @@ export class WeatherForecastDataService {
   }
   getDefaultWeatherForecast(): WeatherForecast {
     return initializeWeatherForecast();
-  }
-
-  getWeatherForecastPerDay(): Observable<ForecastDataPerDay> {
-    return this.forecastPerDaySubject.asObservable();
-  }
-
-  getUnits(): Observable<Units> {
-    return this.units.asObservable();
   }
 }
